requirements.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==2.15.4
cachetools==5.5.2
caer==2.0.8
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
dill==0.3.6
distro==1.9.0
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
isort==5.12.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8
langsmith==0.3.45
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy==1.3.0
mypy-extensions==1.0.0
numpy==1.24.3
oauthlib==3.2.2
openai==1.86.0
opencv-contrib-python==4.7.0.72
orjson==3.10.18
packaging==24.2
platformdirs==3.5.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pylint==2.17.4
pyparsing==3.2.3
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
tomlkit==0.11.8
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
wrapt==1.15.0
zstandard==0.23.0
